ai_service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.9",
    "langchain-google-genai>=2.1.4",
    "langchain-groq>=0.3.2",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.1",
]
